requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.8.0
python-dotenv==1.0.0

openai==1.3.5
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
supabase>=2.0.0

speech-recognition==3.10.0
elevenlabs==0.2.26
python-dateutil==2.8.2
pytz==2023.3

requests==2.31.0
python-multipart==0.0.6
aiofiles==23.2.1

pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
streamlit==1.28.0